feat: clean crud for media

+++ omoide/use_cases/api/uc_media.py
@@ -5,7 +5,8 @@ import base64
 from uuid import UUID
 
 from omoide import domain, utils
-from omoide.domain import interfaces, exceptions
+from omoide.domain import interfaces, errors, actions
+from omoide.infra.special_types import Result, Failure, Success
 from omoide.presentation import api_models
 
 __all__ = [
@@ -20,33 +21,10 @@ class BaseMediaUseCase:
 
     def __init__(
             self,
-            items_repo: interfaces.AbsItemsRepository,
             media_repo: interfaces.AbsMediaRepository,
     ) -> None:
         """Initialize instance."""
-        self.items_repo = items_repo
         self.media_repo = media_repo
-
-    async def _assert_has_access(
-            self,
-            user: domain.User,
-            uuid: UUID,
-    ) -> domain.AccessStatus:
-        """Raise if user has no access to this Media."""
-        access = await self.items_repo.check_access(user, uuid)
-
-        if access.does_not_exist:
-            raise exceptions.NotFound(f'Item {uuid} does not exist')
-
-        if access.is_not_given:
-            raise exceptions.Forbidden(f'User {user.uuid} ({user.name}) '
-                                       f'has no access to item {uuid}')
-
-        if access.is_not_owner:
-            raise exceptions.Forbidden(f'You must own item {uuid} '
-                                       'to be able to modify it')
-
-        return access
 
 
 class CreateOrUpdateMediaUseCase(BaseMediaUseCase):
@@ -61,25 +39,33 @@ class CreateOrUpdateMediaUseCase(BaseMediaUseCase):
 
     async def execute(
             self,
+            policy: interfaces.AbsPolicy,
             user: domain.User,
             uuid: UUID,
             media_type: str,
             media: api_models.CreateMediaIn,
-    ) -> bool:
+    ) -> Result[errors.Error, bool]:
         """Business logic."""
-        await self._assert_has_access(user, uuid)
+        async with self.media_repo.transaction():
+            error = await policy.is_restricted(user, uuid,
+                                               actions.Media.CREATE_OR_UPDATE)
 
-        valid_media = domain.Media(
-            item_uuid=uuid,
-            created_at=utils.now(),
-            processed_at=None,
-            status='init',
-            content=self.extract_binary_content(media.content),
-            ext=media.ext,
-            media_type=media_type,
-        )
+            if error:
+                return Failure(error)
 
-        return await self.media_repo.create_or_update_media(user, valid_media)
+            valid_media = domain.Media(
+                item_uuid=uuid,
+                created_at=utils.now(),
+                processed_at=None,
+                status='init',
+                content=self.extract_binary_content(media.content),
+                ext=media.ext,
+                media_type=media_type,
+            )
+
+            created = await self.media_repo.create_or_update_media(user,
+                                                                   valid_media)
+        return Success(created)
 
 
 class ReadMediaUseCase(BaseMediaUseCase):
@@ -87,15 +73,21 @@ class ReadMediaUseCase(BaseMediaUseCase):
 
     async def execute(
             self,
+            policy: interfaces.AbsPolicy,
             user: domain.User,
             uuid: UUID,
             media_type: str,
-    ) -> domain.Media:
-        await self._assert_has_access(user, uuid)
+    ) -> Result[errors.Error, domain.Media]:
+        async with self.media_repo.transaction():
+            error = await policy.is_restricted(user, uuid, actions.Media.READ)
+
+        if error:
+            return Failure(error)
+
         media = await self.media_repo.read_media(uuid, media_type)
 
         if media is None:
-            raise exceptions.NotFound(f'Media {uuid} does not exist')
+            return Failure(errors.MediaDoesNotExist(uuid=uuid))
 
         return media
 
@@ -105,15 +97,22 @@ class DeleteMediaUseCase(BaseMediaUseCase):
 
     async def execute(
             self,
+            policy: interfaces.AbsPolicy,
             user: domain.User,
             uuid: UUID,
             media_type: str,
-    ) -> bool:
+    ) -> Result[errors.Error, bool]:
         """Business logic."""
-        await self._assert_has_access(user, uuid)
-        deleted = await self.media_repo.delete_media(uuid, media_type)
+        async with self.media_repo.transaction():
+            error = await policy.is_restricted(user, uuid,
+                                               actions.Media.DELETE)
 
-        if not deleted:
-            raise exceptions.NotFound(f'Media {uuid} does not exist')
+            if error:
+                return Failure(error)
+
+            deleted = await self.media_repo.delete_media(uuid, media_type)
+
+            if not deleted:
+                return Failure(errors.MediaDoesNotExist(uuid=uuid))
 
         return True
